Change how the decoder is run in the background
The previous strategy (starting a web worker in the content script)
fails on some sites (e.g. GitHub) with more restrictive content security
policies. A new strategy was needed.

For Firefox, we can just pass the data to and from the background
script; however, this relies on being able to send `Uint8Array`s, which
isn't possible in Chrome since messages must be JSON-serializable.
Passing data through JSON arrays or base64 strings proved to be too
slow, and fails if the data is too large. However, we *can* convert the
data into object URLs and fetch them in the background. This requires a
little extra setup, since Chrome service workers don't have access to
`URL.createObjectURL()`, but it's still possible by doing the decoding
in an offscreen document instead.